added google auth

+++ components/Prayer.jsx
@@ -21,7 +21,7 @@ const Prayer = ({address, id, prayer, timestamp}) => {
     const [comment, setComment] = useState('')
     const [comments, setComments] = useState([])
     const router = useRouter()
-    const { data: session, status } = useSession()
+    const { data: session } = useSession()
 
     useEffect( 
         () => 
@@ -38,7 +38,7 @@ const Prayer = ({address, id, prayer, timestamp}) => {
     [likes])
     
     const likepost = async () => {
-        if(!isAuthenticated) return
+        if(!isAuthenticated || !session) return
         if(hasliked) {
             await deleteDoc(doc(db, 'Prayers', id, 'likes', userAddress || session?.user?.email))
         } else {
@@ -110,7 +110,7 @@ const Prayer = ({address, id, prayer, timestamp}) => {
             <div className='flex items-center p-1'>
                 <BsEmojiSmile size={25} />
                 <input disabled={!session || !isAuthenticated} required maxLength={60} minLength={1} type='text' value={comment} onChange={(e) => {setComment(e.target.value)}} placeholder={` ${isAuthenticated || session ? 'Add a comment...' : 'login to comment'}`} className='bg-[#e4e6e8] p-2 w-full border-none flex-1 focus:ring-0 outline-none' />
-                <button type='submit' disabled={!comment.trim()} onClick={sendComment} className={`${!comment || !isAuthenticated || !session ? 'bg-[#888] text-[#fff] cursor-not-allowed' : ' text-white bg-[#16bafb] cursor-pointer transition-all ease-out duration-300 hover:scale-105' } rounded-[50%] p-[5px] flex  `}>
+                <button type='submit' disabled={!comment.trim()} onClick={sendComment} className={`${!comment || !isAuthenticated || !session ?  'bg-[#888] text-[#fff] cursor-not-allowed' : ' text-white bg-[#16bafb] cursor-pointer transition-all ease-out duration-300 hover:scale-105' } rounded-[50%] p-[5px] flex  `}>
                     <RiSendPlane2Fill  size={25} />
                 </button>            
             </div>
